Split fetch.py into ncbi/metadata/europepmc modules to separate Entrez, enrichment, and Europe PMC concerns (no import cycles)
