+++ src/python/ensembl/io/genomio/literature/fetch.py
@@ -12,233 +12,29 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Retrieve an assembly's publication and supplementary text from NCBI, BioProject and Europe PMC."""
+"""Score candidate papers and orchestrate literature metadata retrieval for an assembly."""
 
-import io
 import logging
-import re as _re
-import time
-import zipfile
-from xml.etree import ElementTree as _ET
 
 import requests
 
+from .metadata import fetch_assembly_metadata, enrich_assembly_metadata
+from .ncbi import (
+    fetch_linked_pubmed_for_assembly,
+    fetch_bioproject_reference_papers,
+    search_ncbi_entrez,
+)
+from .europepmc import (
+    search_by_pmid,
+    search_europe_pmc_by_name,
+    search_europe_pmc_fulltext,
+    get_fulltext_by_pmcid,
+    validate_paper_ids,
+    get_best_available_text,
+)
+
 logger = logging.getLogger(__name__)
 
-# NCBI E-utilities base URL, shared by the Entrez helpers below.
-NCBI_EUTILS = "https://eutils.ncbi.nlm.nih.gov/entrez/eutils"
-
-
-def entrez_get(endpoint: str, params: dict) -> requests.Response:
-    """Call an NCBI E-utilities endpoint (e.g. "esearch.fcgi") and return the raw
-    response. Raises on HTTP error; callers wrap it in their own try/except."""
-    response = requests.get(f"{NCBI_EUTILS}/{endpoint}", params=params, timeout=30)  # type: ignore[arg-type]
-    response.raise_for_status()
-    return response
-
-
-def entrez_json(endpoint: str, params: dict) -> dict:
-    """Call an NCBI E-utilities endpoint and return the parsed JSON response."""
-    return entrez_get(endpoint, params).json()
-
-
-def _entrez_elink(dbfrom: str, db: str, uid: str) -> list:
-    """Return NCBI elink links from `dbfrom` to `db` for one UID ([] on any failure)."""
-    try:
-        data = entrez_json("elink.fcgi", {"dbfrom": dbfrom, "db": db, "id": uid, "retmode": "json"})
-        linksets = data.get("linksets", [{}])[0]
-        for ldb in linksets.get("linksetdbs", []):
-            if ldb.get("dbto") == db:
-                return ldb.get("links", [])
-    except (requests.RequestException, ValueError, KeyError, IndexError):
-        pass
-    return []
-
-
-# ============================================================
-# STEP 1 — Get assembly info from NCBI
-# ============================================================
-
-
-def _fetch_assembly_report(accession: str) -> list:
-    """Query NCBI Datasets for one accession's dataset report; return the raw
-    reports list (empty on any failure). Isolated so callers can retry with a
-    different accession form (e.g. version-less) without duplicating HTTP logic."""
-    url = f"https://api.ncbi.nlm.nih.gov/datasets/v2/genome/accession" f"/{accession}/dataset_report"
-    try:
-        response = requests.get(url, timeout=30)
-        response.raise_for_status()
-        return response.json().get("reports", [])
-    except (requests.RequestException, ValueError) as e:
-        logger.warning(f"  [NCBI] Failed to fetch {accession}: {e}")
-        return []
-
-
-def fetch_assembly_metadata(accession: str) -> dict:
-    """Fetch an assembly's NCBI Datasets metadata (species, taxon_id, common name,
-    linked PMIDs, chromosome number), retrying without a stale version suffix.
-    Returns a dict with empty/None fields if no report is found."""
-    reports = _fetch_assembly_report(accession)
-
-    # A stale version suffix (e.g. GCA_000001405.15 when .29 is current) makes
-    # NCBI return an empty report. A version-less accession resolves to the
-    # latest version, so retry without the ".N" suffix before giving up — this
-    # keeps the pipeline working when accession lists drift out of date.
-    resolved_accession = accession
-    if not reports and "." in accession:
-        base = accession.split(".")[0]
-        logger.info(f"  [NCBI] No report for {accession}; retrying latest version ({base}) ...")
-        reports = _fetch_assembly_report(base)
-
-    if not reports:
-        logger.info(f"  [NCBI] No report found for {accession}")
-        return {
-            "assembly_accession": accession,
-            "chromosome_number": None,
-            "chromosome_source": None,
-        }
-
-    report = reports[0]
-    resolved_accession = report.get("accession") or resolved_accession
-    assembly_info = report.get("assembly_info", {})
-    organism = report.get("organism", {})
-    assembly_stats = report.get("assembly_stats", {})
-
-    chromosome_number = assembly_stats.get("total_number_of_chromosomes") or assembly_info.get(
-        "chromosome_count"
-    )
-
-    raw_pmid = assembly_info.get("linked_pmid") or assembly_info.get("biosample", {}).get("linked_pmid")
-    linked_pmids = raw_pmid if isinstance(raw_pmid, list) else [raw_pmid] if raw_pmid else []
-
-    return {
-        "assembly_accession": resolved_accession,
-        "assembly_name": assembly_info.get("assembly_name"),
-        "taxon_id": str(organism.get("tax_id") or ""),
-        "scientific_name": organism.get("organism_name"),
-        "common_name": organism.get("common_name"),  # NEW — mentor: search by common name too
-        "linked_pmids": linked_pmids,
-        "chromosome_number": chromosome_number,
-        "chromosome_source": "ncbi" if chromosome_number else None,
-    }
-
-
-# ============================================================
-# Taxonomy enrichment (use taxon_id to fill gaps)
-#   - common_name from NCBI Taxonomy efetch (fixes species like maize where
-#     NCBI datasets returns no common_name -> name search degrades)
-#   - reference ploidy / chromosome number from GoaT (taxon-based, sourced)
-# ============================================================
-
-
-def fetch_taxonomy_common_name(taxon_id: str) -> str | None:
-    """Look up a taxon's common name from NCBI Taxonomy (GenBank common name
-    preferred), or None if unavailable."""
-    # Caller (enrich_assembly_metadata) guards on taxon_id before calling this.
-    try:
-        r = entrez_get("efetch.fcgi", {"db": "taxonomy", "id": str(taxon_id)})
-        root = _ET.fromstring(r.text)
-        taxon = root.find("Taxon")
-        if taxon is None:
-            return None
-        # GenbankCommonName preferred, else the first CommonName
-        genbank_cname = taxon.find(".//OtherNames/GenbankCommonName")
-        if genbank_cname is not None and genbank_cname.text:
-            return genbank_cname.text.strip()
-        common_name = taxon.find(".//OtherNames/CommonName")
-        if common_name is not None and common_name.text:
-            return common_name.text.strip()
-    except (requests.RequestException, _ET.ParseError):
-        pass
-    return None
-
-
-def fetch_goat_reference(taxon_id: str) -> dict:
-    """Look up taxon-level reference ploidy / chromosome number from GoaT.
-
-    GoaT stores attributes under records[0].record.attributes. Each attribute
-    carries an `aggregation_source`: "direct" = a directly measured/curated
-    value; "ancestor"/"descendant" = phylogenetically *inferred*. We only trust
-    DIRECT values — inferred ones repeat the polyploid-ancestor trap (e.g. for
-    Arabidopsis, GoaT's ploidy_inferred is 4 via the tribe, which is wrong).
-    Any unexpected shape -> returns Nones so the pipeline falls back safely.
-    """
-    out: dict[str, int | None] = {"reference_ploidy": None, "reference_chromosome": None}
-    # Caller (enrich_assembly_metadata) guards on taxon_id before calling this.
-    url = "https://goat.genomehubs.org/api/v2/record"
-    try:
-        r = requests.get(
-            url,
-            params={
-                "recordId": str(taxon_id),
-                "result": "taxon",
-                "taxonomy": "ncbi",
-            },
-            timeout=30,
-        )
-        r.raise_for_status()
-        recs = r.json().get("records", [])
-        if not recs:
-            return out
-        attrs = (recs[0].get("record", {}) or {}).get("attributes", {}) or {}
-
-        def _direct_int(*keys: str) -> int | None:
-            # try each attribute key in order; accept only directly-measured values
-            for key in keys:
-                attr = attrs.get(key)
-                if not isinstance(attr, dict):
-                    continue
-                if attr.get("aggregation_source") != "direct":
-                    continue  # skip inferred (ancestor/descendant) values
-                for value_key in ("value", "median", "mode", "max", "min"):
-                    value = attr.get(value_key)
-                    if value is None:
-                        continue
-                    match = _re.search(r"\d+", str(value))
-                    if match:
-                        return int(match.group())
-            return None
-
-        out["reference_ploidy"] = _direct_int("ploidy", "ploidy_inferred")
-        out["reference_chromosome"] = _direct_int("chromosome_number", "haploid_chromosome_count")
-    except (requests.RequestException, ValueError, KeyError, TypeError):
-        pass
-    return out
-
-
-def enrich_assembly_metadata(assembly: dict) -> dict:
-    """Fill common_name (if missing) and attach GoaT reference values, using taxon_id."""
-    taxon_id = assembly.get("taxon_id")
-    if not taxon_id:
-        return assembly
-
-    if not assembly.get("common_name"):
-        common_name = fetch_taxonomy_common_name(taxon_id)
-        if common_name:
-            assembly["common_name"] = common_name
-            assembly["common_name_source"] = "ncbi_taxonomy"
-            logger.info(f"      common_name filled from NCBI Taxonomy: {common_name}")
-
-    goat = fetch_goat_reference(taxon_id)
-    assembly["reference_ploidy"] = goat.get("reference_ploidy")
-    assembly["reference_chromosome"] = goat.get("reference_chromosome")
-    if goat.get("reference_ploidy") is not None:
-        logger.info(f"      GoaT reference ploidy (taxon {taxon_id}): {goat['reference_ploidy']}")
-    return assembly
-
-
-# ============================================================
-# STEP 2 — Find linked papers
-#           Strategy (Mode B — escalate on weak result):
-#           1st: directly linked PMIDs from NCBI assembly record
-#           2nd: NCBI Entrez elink (curated assembly -> pubmed links)
-#           -- if no STRONG genome-paper candidate from the above, escalate:
-#           3rd: Europe PMC search by scientific_name AND common_name
-#           4th: NCBI Entrez PMC search
-#           5th: Semantic Scholar
-#           Candidates from all sources are POOLED, de-duplicated, and
-#           ranked best-first by score_paper_candidate().
-# ============================================================
 
 ORGANELLE_KEYWORDS = [
     "chloroplast",
@@ -249,6 +45,8 @@ ORGANELLE_KEYWORDS = [
     "mtdna",
     "plastome",
 ]
+
+
 GENOME_KEYWORDS = [
     "genome assembly",
     "genome sequence",
@@ -260,6 +58,8 @@ GENOME_KEYWORDS = [
     "whole-genome",
     "genome of",
 ]
+
+
 # Phrases that mark the target organism as a *habitat/host*, not the subject
 # of the paper — e.g. "Draft genome of Bacillus ... isolated from cotton". Such
 # papers mention the target's name only as the source and otherwise score high
@@ -278,6 +78,8 @@ HOST_CONTEXT_KEYWORDS = [
     "gut of",
     "gut microb",
 ]
+
+
 PLOIDY_HINTS = [
     "diploid",
     "triploid",
@@ -293,6 +95,7 @@ PLOIDY_HINTS = [
     "2n=",
     "2n =",
 ]
+
 
 # A candidate scoring >= STRONG_SCORE is treated as a confident genome paper,
 # so Mode B will NOT escalate to extra name-based searches.
@@ -396,142 +199,6 @@ def _dedupe_and_rank(candidates: list, identifiers: dict) -> list:
     return rank
 
 
-def search_europe_pmc(query: str, max_results: int = 5, retries: int = 3) -> list:
-    """Search Europe PMC for `query` and return up to `max_results` paper records,
-    retrying transient request failures up to `retries` times."""
-    url = "https://www.ebi.ac.uk/europepmc/webservices/rest/search"
-    params = {
-        "query": query,
-        "format": "json",
-        "pageSize": max_results,
-        "resultType": "core",
-    }
-    for attempt in range(retries):
-        try:
-            response = requests.get(url, params=params, timeout=30)  # type: ignore[arg-type]
-            response.raise_for_status()
-            return response.json().get("resultList", {}).get("result", [])
-        except requests.RequestException as e:
-            if attempt < retries - 1:
-                wait = 2**attempt
-                logger.warning(f"  [EuropePMC] Attempt {attempt+1} failed, retrying in {wait}s ...")
-                time.sleep(wait)
-            else:
-                logger.warning(f"  [EuropePMC] All {retries} attempts failed: {e}")
-                return []
-    return []
-
-
-def search_by_pmid(pmid: str) -> dict | None:
-    """Fetch one paper's Europe PMC record by PMID, or None if not found."""
-    url = "https://www.ebi.ac.uk/europepmc/webservices/rest/search"
-    params = {
-        "query": f"EXT_ID:{pmid} AND SRC:MED",
-        "format": "json",
-        "resultType": "core",
-    }
-    try:
-        response = requests.get(url, params=params, timeout=30)
-        response.raise_for_status()
-        results = response.json().get("resultList", {}).get("result", [])
-        return results[0] if results else None
-    except (requests.RequestException, ValueError):
-        return None
-
-
-def search_europe_pmc_by_name(
-    taxon_id: str,
-    scientific_name: str,
-    common_name: str = "",
-    max_results: int = 5,
-) -> list:
-    """Search Europe PMC for genome papers by scientific/common name, escalating
-    from full name to binomial to genus to taxon-ID queries and keeping only
-    genus-relevant hits. Returns up to max_results records ([] if none)."""
-    tokens = scientific_name.split()  # "".split() -> [], so no guard needed
-    genus = tokens[0] if tokens else ""
-    # Binomial = genus + species epithet. For a subspecies / variety / hybrid
-    # (e.g. "Oryza meyeriana var. indandamanica") the full trinomial rarely
-    # appears in a paper, but the binomial ("Oryza meyeriana") usually does.
-    binomial = " ".join(tokens[:2])
-
-    # Multi-identifier name clause: scientific name, plus common name when
-    # it is specific enough to be useful (>= 4 chars avoids junk like "fly").
-    name_clause = f'"{scientific_name}"'
-    if common_name and len(common_name) >= 4:
-        name_clause = f'("{scientific_name}" OR "{common_name}")'
-
-    queries = [
-        # 1st priority: name(s) + genome assembly
-        f'{name_clause} AND "genome assembly"',
-        # 2nd priority: name(s) + genome sequence / ploidy / chromosome
-        f'{name_clause} AND ("genome sequence" OR ploidy OR chromosome)',
-        # 3rd priority: scientific name + genome sequence
-        f'"{scientific_name}" AND "genome sequence"',
-    ]
-
-    # 4th priority: taxonomic fallback. When the full name is a trinomial
-    # (subspecies / variety / hybrid), drop to the binomial, then the genus,
-    # before the taxon-ID last resort. The genus relevance filter below still
-    # applies, so this widens recall without inviting unrelated papers.
-    if binomial and binomial != scientific_name:
-        queries += [
-            f'"{binomial}" AND "genome assembly"',
-            f'"{binomial}" AND ("genome sequence" OR ploidy OR chromosome)',
-        ]
-
-    if taxon_id:
-        queries += [
-            # 5th priority: taxon ID fallback (last resort)
-            f"TAXONOMY:{taxon_id} AND genome assembly",
-        ]
-
-    for query in queries:
-        results = search_europe_pmc(query, max_results=max_results * 3)
-        if not results:
-            continue
-
-        # Keep only results that actually mention the genus — guards against
-        # false positives from an ambiguous common name.
-        relevant = [
-            r for r in results if genus.lower() in (r.get("title", "") + r.get("abstractText", "")).lower()
-        ]
-
-        if relevant:
-            logger.info(f"  [EuropePMC] Query: {query}")
-            logger.info(
-                f"              Found {len(results)} results, "
-                f"{len(relevant)} relevant to {scientific_name}"
-            )
-            return relevant[:max_results]
-
-    logger.info(f"  [EuropePMC] No relevant papers found for {scientific_name}")
-    return []
-
-
-def search_europe_pmc_fulltext(scientific_name: str, max_results: int = 5) -> list:
-    """Find candidate source papers that mention the exact species binomial in
-    their FULL TEXT even when the title/abstract does not — e.g. a multi-species
-    or methods paper that assembled this genome (the true source of many
-    assemblies). Europe PMC's search covers open-access full text, so a quoted
-    binomial query surfaces them. We deliberately skip the genus-in-title guard
-    here (that guard drops exactly these papers) and instead return only
-    open-access hits, tagged for a full-text confirmation pass in the caller."""
-    if not scientific_name:
-        return []
-    query = f'"{scientific_name}" AND (genome OR assembly OR chromosome OR ploidy)'
-    candidates = []
-    for paper in search_europe_pmc(query, max_results=max_results * 3):
-        if paper.get("pmcid") or paper.get("pmCid"):
-            paper["_needs_fulltext_confirm"] = True
-            candidates.append(paper)
-        if len(candidates) >= max_results * 2:
-            break
-    if candidates:
-        logger.info(f"  [EuropePMC] Full-text species search: {len(candidates)} open-access candidate(s)")
-    return candidates
-
-
 def has_usable_fulltext(papers: list[dict]) -> bool:
     """Return True if at least one paper has a PMCID (open-access full text) and
     is not an organelle (chloroplast/mitochondrion) genome paper."""
@@ -545,147 +212,6 @@ def has_usable_fulltext(papers: list[dict]) -> bool:
         if not is_organelle:
             return True
     return False
-
-
-def fetch_linked_pubmed_for_assembly(accession: str, max_results: int = 5) -> list:
-    """Follow the NCBI assembly -> PubMed elink to retrieve linked publications.
-    Returns Europe PMC paper records ([] if no assembly UID or no links found)."""
-    # Step 1: convert GCA accession to NCBI assembly UID
-    try:
-        data = entrez_json("esearch.fcgi", {"db": "assembly", "term": accession, "retmode": "json"})
-        assembly_ids = data.get("esearchresult", {}).get("idlist", [])
-    except (requests.RequestException, ValueError) as e:
-        logger.warning(f"  [NCBI Entrez elink] Assembly search failed: {e}")
-        return []
-
-    if not assembly_ids:
-        logger.info(f"  [NCBI Entrez elink] No assembly ID found for {accession}")
-        return []
-
-    assembly_id = assembly_ids[0]
-    logger.info(f"  [NCBI Entrez elink] Assembly ID: {assembly_id}")
-
-    # Step 2: find linked PubMed articles via elink
-    pmids = _entrez_elink("assembly", "pubmed", assembly_id)
-
-    if not pmids:
-        logger.info(f"  [NCBI Entrez elink] No linked publications found")
-        return []
-
-    logger.info(f"  [NCBI Entrez elink] Found {len(pmids)} linked PMID(s): {pmids[:5]}")
-
-    # Step 3: fetch paper details from Europe PMC using PMIDs
-    papers = []
-    for pmid in pmids[:max_results]:
-        paper = search_by_pmid(str(pmid))
-        if paper:
-            papers.append(paper)
-            logger.info(f"  [NCBI Entrez elink] {pmid}: {paper.get('title', '')[:60]}...")
-
-    return papers
-
-
-def fetch_bioproject_reference_papers(accession: str, max_results: int = 5) -> list:
-    """Mentor item 3: follow assembly -> BioProject -> PubMed to retrieve the
-    canonical *reference* paper for the assembly. These are attached even when
-    they carry no ploidy information (a BioProject paper still anchors species /
-    common name / provenance). Pure-Entrez chain; defensive on every step."""
-    # Step 1: accession -> assembly UID
-    try:
-        data = entrez_json("esearch.fcgi", {"db": "assembly", "term": accession, "retmode": "json"})
-        asm_ids = data.get("esearchresult", {}).get("idlist", [])
-    except (requests.RequestException, ValueError):
-        asm_ids = []
-    if not asm_ids:
-        return []
-
-    # Step 2: assembly UID -> BioProject UID(s)
-    bioproject_ids = _entrez_elink("assembly", "bioproject", asm_ids[0])
-    if not bioproject_ids:
-        logger.info("  [BioProject] No linked BioProject found")
-        return []
-    logger.info(f"  [BioProject] Linked BioProject UID(s): {bioproject_ids[:5]}")
-
-    # Step 3: BioProject UID(s) -> reference PMIDs
-    pmids = []
-    for bp in bioproject_ids[:3]:
-        for pmid in _entrez_elink("bioproject", "pubmed", bp):
-            if pmid not in pmids:
-                pmids.append(pmid)
-    if not pmids:
-        logger.info("  [BioProject] No reference publication linked to BioProject")
-        return []
-    logger.info(f"  [BioProject] Found {len(pmids)} reference PMID(s): {pmids[:5]}")
-
-    # Step 4: fetch paper details, tag as reference paper
-    papers = []
-    for pmid in pmids[:max_results]:
-        paper = search_by_pmid(str(pmid))
-        if paper:
-            paper["retrieval_source"] = "bioproject"
-            paper["is_reference_paper"] = True
-            papers.append(paper)
-            logger.info(f"  [BioProject] {pmid}: {paper.get('title', '')[:60]}...")
-    return papers
-
-
-def search_ncbi_entrez(scientific_name: str, max_results: int = 5) -> list:
-    """NCBI Entrez PMC search fallback: query PubMed Central for open-access
-    genome-assembly papers by scientific name. Used when accession links and the
-    Europe PMC name search fail to surface a strong genome paper."""
-    logger.info(f"  [NCBI Entrez] Searching PubMed Central for {scientific_name} ...")
-
-    # Step 1: search for PMC IDs
-    try:
-        data = entrez_json(
-            "esearch.fcgi",
-            {
-                "db": "pmc",
-                "term": f'"{scientific_name}" AND "genome assembly" AND "open access"[filter]',
-                "retmax": max_results * 2,
-                "retmode": "json",
-            },
-        )
-        pmc_ids = data.get("esearchresult", {}).get("idlist", [])
-    except (requests.RequestException, ValueError) as e:
-        logger.warning(f"  [NCBI Entrez] Search failed: {e}")
-        return []
-
-    if not pmc_ids:
-        logger.info(f"  [NCBI Entrez] No PMC IDs found")
-        return []
-
-    # Step 2: fetch summaries
-    try:
-        data = entrez_json(
-            "esummary.fcgi",
-            {"db": "pmc", "id": ",".join(pmc_ids[:max_results]), "retmode": "json"},
-        )
-        summaries = data.get("result", {})
-    except (requests.RequestException, ValueError) as e:
-        logger.warning(f"  [NCBI Entrez] Summary fetch failed: {e}")
-        return []
-
-    # Step 3: convert to Europe PMC format
-    papers = []
-    for pmc_id in pmc_ids[:max_results]:
-        summary = summaries.get(pmc_id, {})
-        if not summary:
-            continue
-        pmcid = f"PMC{pmc_id}"
-        papers.append(
-            {
-                "title": summary.get("title", ""),
-                "pmcid": pmcid,
-                "pmid": summary.get("pmid", ""),
-                "doi": summary.get("doi", ""),
-                "abstractText": "",
-                "source": "ncbi_entrez",
-            }
-        )
-        logger.info(f"  [NCBI Entrez] {pmcid}: {summary.get('title', '')[:60]}...")
-
-    return papers
 
 
 def search_semantic_scholar(scientific_name: str, max_results: int = 5) -> list:
@@ -735,199 +261,6 @@ def search_semantic_scholar(scientific_name: str, max_results: int = 5) -> list:
 
     logger.info(f"  [Semantic Scholar] Found {len(papers)} relevant paper(s)")
     return papers
-
-
-# ============================================================
-# STEP 3 — Validate DOI / PMID / PMCID
-# ============================================================
-
-
-def validate_paper_ids(paper: dict) -> dict:
-    """Check a paper's PMCID / PMID against Europe PMC and report which text tier
-    (full text, abstract, or none) is actually retrievable for it."""
-    pmcid = paper.get("pmcid")
-    pmid = paper.get("pmid")
-    doi = paper.get("doi")
-
-    pmcid_valid = False
-    if pmcid:
-        try:
-            url = f"https://www.ebi.ac.uk/europepmc/webservices/rest/{pmcid}/fullTextXML"
-            r = requests.head(url, timeout=10)
-            pmcid_valid = r.status_code == 200
-        except requests.RequestException:
-            pmcid_valid = False
-
-    pmid_valid = False
-    if pmid:
-        try:
-            result = search_by_pmid(pmid)
-            pmid_valid = result is not None
-        except requests.RequestException:
-            pmid_valid = False
-
-    return {
-        "pmcid": pmcid,
-        "pmid": pmid,
-        "doi": doi,
-        "pmcid_valid": pmcid_valid,
-        "pmid_valid": pmid_valid,
-        "doi_present": bool(doi),
-        "has_fulltext": pmcid_valid,
-        "best_available": ("fulltext" if pmcid_valid else "abstract" if pmid_valid else "none"),
-    }
-
-
-# ============================================================
-# STEP 4 — Download full text or abstract
-# ============================================================
-
-
-def get_fulltext_by_pmcid(pmcid: str) -> str | None:
-    """Download an article's full-text XML from Europe PMC by PMCID, or None."""
-    url = f"https://www.ebi.ac.uk/europepmc/webservices/rest/{pmcid}/fullTextXML"
-    try:
-        response = requests.get(url, timeout=30)
-        if response.status_code == 200:
-            return response.text
-    except requests.RequestException:
-        pass
-    return None
-
-
-def _strip_markup(s: str) -> str:
-    try:
-        from bs4 import BeautifulSoup
-
-        return BeautifulSoup(s, "html.parser").get_text(" ")
-    except Exception:
-        return _re.sub(r"<[^>]+>", " ", s)
-
-
-def _read_zip(raw: bytes, member: str) -> str:
-    """Read one member from an in-memory zip (xlsx/docx are zip archives) and
-    return its decoded text, or '' if the archive or the member is unreadable."""
-    try:
-        archive = zipfile.ZipFile(io.BytesIO(raw))
-        if member not in archive.namelist():
-            return ""
-        return archive.read(member).decode("utf-8", errors="ignore")
-    except Exception:
-        return ""
-
-
-def _xlsx_shared_strings(raw: bytes) -> str:
-    """Extract shared-string text from an .xlsx (itself a zip) without openpyxl.
-    This catches column headers / row labels ('2n', 'chromosome number', ploidy
-    words) that carry the ploidy signal; numeric cell values are not needed for
-    that. Returns '' if the file is not a readable xlsx."""
-    xml = _read_zip(raw, "xl/sharedStrings.xml")
-    return " ".join(_re.findall(r"<t[^>]*>([^<]+)</t>", xml))
-
-
-def _docx_text(raw: bytes) -> str:
-    """Extract paragraph text from a .docx (itself a zip) without python-docx:
-    read word/document.xml and pull the <w:t> text runs. Supplementary tables
-    are very commonly distributed as .docx, so this materially widens coverage.
-    Returns '' if the file is not a readable docx."""
-    xml = _read_zip(raw, "word/document.xml")
-    return " ".join(_re.findall(r"<w:t[^>]*>([^<]*)</w:t>", xml))
-
-
-def get_supplementary_text_by_pmcid(pmcid: str, max_chars: int = 40000) -> str | None:
-    """Download the Europe PMC supplementary-files ZIP for a PMC article and
-    extract plain text from the members we can read without heavy dependencies:
-    txt/csv/tsv, html/xml, and xlsx shared strings. Ploidy evidence —
-    chromosome counts, karyotype tables, '2n = ...' formulas — often lives in
-    supplementary tables rather than the article body, so this widens what the
-    extractor can see. Binary formats we cannot read dependency-free (pdf, docx,
-    images) are skipped. Returns one concatenated string, or None if nothing
-    usable was found."""
-    url = f"https://www.ebi.ac.uk/europepmc/webservices/rest/{pmcid}/supplementaryFiles"
-    try:
-        resp = requests.get(url, timeout=60)
-        if resp.status_code != 200 or not resp.content:
-            return None
-    except requests.RequestException:
-        return None
-
-    try:
-        zf = zipfile.ZipFile(io.BytesIO(resp.content))
-    except zipfile.BadZipFile:
-        return None
-
-    parts, total = [], 0
-    for name in zf.namelist():
-        if total >= max_chars:
-            break
-        lower = name.lower()
-        try:
-            raw = zf.read(name)
-        except Exception:
-            continue
-        if lower.endswith((".txt", ".csv", ".tsv", ".tab")):
-            text = raw.decode("utf-8", errors="ignore")
-        elif lower.endswith((".html", ".htm", ".xml")):
-            text = _strip_markup(raw.decode("utf-8", errors="ignore"))
-        elif lower.endswith(".xlsx"):
-            text = _xlsx_shared_strings(raw)
-        elif lower.endswith(".docx"):
-            text = _docx_text(raw)
-        else:
-            continue  # pdf / legacy .doc / images: not readable dependency-free
-        text = _re.sub(r"[ \t\r\n\f\v]+", " ", text).strip()
-        if text:
-            snippet = text[: max_chars - total]
-            parts.append(f"[Supplementary file: {name}]\n{snippet}")
-            total += len(snippet)
-
-    if not parts:
-        return None
-    logger.info(f"  [supp] extracted text from {len(parts)} supplementary file(s) for {pmcid}")
-    return "\n\n".join(parts)
-
-
-def get_abstract_by_pmid(pmid: str) -> str | None:
-    """Return the abstract text for a PMID via Europe PMC, or None."""
-    paper = search_by_pmid(pmid)
-    if paper:
-        return paper.get("abstractText")
-    return None
-
-
-def get_best_available_text(paper: dict) -> dict:
-    """Return the best retrievable text for a paper as {source, text, ...}: full
-    text (+ supplementary) by PMCID, else abstract by PMID, else the search-result
-    abstract, else source 'none'."""
-    pmcid = paper.get("pmcid")
-    pmid = paper.get("pmid")
-
-    if pmcid:
-        fulltext = get_fulltext_by_pmcid(pmcid)
-        if fulltext:
-            out = {"source": "fulltext", "pmcid": pmcid, "text": fulltext}
-            supp = get_supplementary_text_by_pmcid(pmcid)
-            if supp:
-                out["supplementary"] = supp
-            return out
-
-    if pmid:
-        abstract = get_abstract_by_pmid(pmid)
-        if abstract:
-            return {"source": "abstract", "pmid": pmid, "text": abstract}
-
-    abstract = paper.get("abstractText")
-    if abstract:
-        return {"source": "search_result", "text": abstract}
-
-    return {"source": "none", "text": None}
-
-
-# ============================================================
-# MAIN FUNCTION — run all steps above in order
-#           input:  assembly accession (e.g. 'GCA_003086295.3')
-#           output: assembly metadata + linked papers with full text
-# ============================================================
 
 
 def _strongest_score(candidates: list, identifiers: dict) -> float:
